+++ climpred/classes.py
@@ -195,13 +195,10 @@ class PerfectModelEnsemble(PredictionEnsemble):
             self.uninitialized, self.control, metric=metric, comparison=comparison
         )
 
-    def compute_persistence(self, nlags=None, metric='pearson_r'):
+    def compute_persistence(self, metric='pearson_r'):
         """Compute a simple persistence forecast for the control run.
 
         Args:
-            nlags (int, default None):
-              Number of lags to compute persistence forecast to. If None,
-              compute to the length of the initialized forecasts.
             metric (str, default 'pearson_r'):
               Metric to apply to the persistence forecast.
 
@@ -216,8 +213,6 @@ class PerfectModelEnsemble(PredictionEnsemble):
               prediction. Oxford University Press, 2007.
         """
         is_initialized(self.control, 'control', 'a persistence forecast')
-        if nlags is None:
-            nlags = self.initialized.lead.size
         return compute_persistence(self.initialized, self.control, metric=metric)
 
     def bootstrap(
@@ -408,11 +403,13 @@ class HindcastEnsemble(PredictionEnsemble):
         """
         if isinstance(xobj, xr.DataArray):
             xobj = xobj.to_dataset()
-        match_initialized_dims(self.initialized, xobj)
+        match_initialized_dims(self.initialized, xobj, uninitialized=True)
         match_initialized_vars(self.initialized, xobj)
         self.uninitialized = xobj
 
-    def compute_metric(self, refname=None, metric='pearson_r', comparison='e2r'):
+    def compute_metric(
+        self, refname=None, metric='pearson_r', comparison='e2r', max_dof=False
+    ):
         """Compares the initialized ensemble to a given reference.
 
         This will automatically run the comparison against all shared variables
@@ -427,6 +424,8 @@ class HindcastEnsemble(PredictionEnsemble):
             comparison (str, default 'e2r'):
               How to compare to the reference. ('e2r' for ensemble mean to
               reference. 'm2r' for each individual member to reference)
+            max_dof (bool, default False):
+              If True, maximize the degrees of freedom for each lag calculation.
 
         Returns:
             Dataset of comparison results (if comparing to one reference),
@@ -442,6 +441,7 @@ class HindcastEnsemble(PredictionEnsemble):
                 self.reference[refname].drop(drop_ref),
                 metric=metric,
                 comparison=comparison,
+                max_dof=max_dof,
             )
         else:
             if len(self.reference) == 1:
@@ -452,6 +452,7 @@ class HindcastEnsemble(PredictionEnsemble):
                     self.reference[refname].drop(drop_ref),
                     metric=metric,
                     comparison=comparison,
+                    max_dof=max_dof,
                 )
             # Loop through all references and return results as a dictionary
             # with keys corresponding to reference names.
@@ -464,6 +465,7 @@ class HindcastEnsemble(PredictionEnsemble):
                         self.reference[key].drop(drop_ref),
                         metric=metric,
                         comparison=comparison,
+                        max_dof=max_dof,
                     )
                 return skill
 
@@ -523,7 +525,7 @@ class HindcastEnsemble(PredictionEnsemble):
                     )
                 return u
 
-    def compute_persistence(self, refname=None, nlags=None, metric='pearson_r'):
+    def compute_persistence(self, refname=None, metric='pearson_r', max_dof=False):
         """Compute a simple persistence forecast for a reference.
 
         This simply applies some metric between the reference and itself out
@@ -533,11 +535,10 @@ class HindcastEnsemble(PredictionEnsemble):
             refname (str, default None):
               Name of reference to compute the persistence forecast for. If
               `None`, compute for all references.
-            nlags (int, default None):
-              Number of lags to compute persistence forecast to. If None,
-              compute to the length of the initialized forecasts.
             metric (str, default 'pearson_r'):
               Metric to apply to the persistence forecast.
+            max_dof (bool, default False):
+              If True, maximize the degrees of freedom for each lag calculation.
 
         Returns:
             Dataset of persistence forecast results (if refname is declared),
@@ -550,19 +551,31 @@ class HindcastEnsemble(PredictionEnsemble):
               prediction. Oxford University Press, 2007.
         """
         is_initialized(self.reference, 'reference', 'a persistence forecast')
-        # Default to the length of the initialized forecast.
-        if nlags is None:
-            nlags = self.initialized.lead.size
         # apply to single reference.
         if refname is not None:
             return compute_persistence(
-                self.initialized, self.reference[refname], metric=metric
+                self.initialized,
+                self.reference[refname],
+                metric=metric,
+                max_dof=max_dof,
             )
         # loop through and apply to all references.
         else:
-            persistence = {}
-            for key in self.reference:
-                persistence[key] = compute_persistence(
-                    self.initialized, self.reference[key], metric=metric
+            if len(self.reference) == 1:
+                refname = list(self.reference.keys())[0]
+                return compute_persistence(
+                    self.initialized,
+                    self.reference[refname],
+                    metric=metric,
+                    max_dof=max_dof,
                 )
+            else:
+                persistence = {}
+                for key in self.reference:
+                    persistence[key] = compute_persistence(
+                        self.initialized,
+                        self.reference[key],
+                        metric=metric,
+                        max_dof=max_dof,
+                    )
             return persistence