+++ climpred/prediction.py
@@ -50,7 +50,7 @@ def compute_perfect_model(ds, control, metric='pearson_r', comparison='m2e'):
 
 @is_xarray([0, 1])
 def compute_hindcast(
-    hind, reference, metric='pearson_r', comparison='e2r', max_dfs=False
+    hind, reference, metric='pearson_r', comparison='e2r', max_dof=False
 ):
     """Compute a predictability skill score against a reference
 
@@ -74,7 +74,7 @@ def compute_hindcast(
                 * m2r : each member to the reference
         nlags (int): How many lags to compute skill/potential predictability out
                      to. Default: length of `lead` dim
-        max_dfs (bool):
+        max_dof (bool):
             If True, maximize the degrees of freedom by slicing `hind` and `reference`
             to a common time frame at each lead.
 
@@ -95,13 +95,13 @@ def compute_hindcast(
     # think in real time dimension: real time = init + lag
     forecast = forecast.rename({'init': 'time'})
     # take only inits for which we have references at all leahind
-    if not max_dfs:
+    if not max_dof:
         forecast, reference = reduce_time_series(forecast, reference, nlags)
 
     plag = []
     # iterate over all leads (accounts for lead.min() in [0,1])
     for i in forecast.lead.values:
-        if max_dfs:
+        if max_dof:
             forecast, reference = reduce_time_series(forecast, reference, i)
         # take lead year i timeseries and convert to real time
         a = forecast.sel(lead=i).drop('lead')
@@ -115,7 +115,7 @@ def compute_hindcast(
 
 
 @is_xarray([0, 1])
-def compute_persistence(hind, reference, metric='pearson_r', max_dfs=False):
+def compute_persistence(hind, reference, metric='pearson_r', max_dof=False):
     """Computes the skill of a persistence forecast from a simulation.
 
     Args:
@@ -123,7 +123,7 @@ def compute_persistence(hind, reference, metric='pearson_r', max_dfs=False):
         reference (xarray object): The reference time series.
         metric (str): Metric name to apply at each lag for the persistence
                       computation. Default: 'pearson_r'
-        max_dfs (bool):
+        max_dof (bool):
             If True, maximize the degrees of freedom by slicing `hind` and `reference`
             to a common time frame at each lead.
 
@@ -150,7 +150,7 @@ def compute_persistence(hind, reference, metric='pearson_r', max_dfs=False):
     nlags = max(hind.lead.values)
     # temporarily change `init` to `time` for comparison to reference time.
     hind = hind.rename({'init': 'time'})
-    if not max_dfs:
+    if not max_dof:
         # slices down to inits in common with hindcast, plus gives enough room
         # for maximum lead time forecast.
         a, _ = reduce_time_series(hind, reference, nlags)
@@ -158,7 +158,7 @@ def compute_persistence(hind, reference, metric='pearson_r', max_dfs=False):
 
     plag = []
     for lag in hind.lead.values:
-        if max_dfs:
+        if max_dof:
             # slices down to inits in common with hindcast, but only gives enough
             # room for lead from current forecast
             a, _ = reduce_time_series(hind, reference, lag)